+++ apps/zui/src/zui-kit/core/index.ts
@@ -1,4 +1,4 @@
 export * from "./list-view/list-view-api"
-export * from "./table-view/table-view-api"
+export * from "../table/table-view-api"
 export * from "./value-view/view-context"
 export * from "./value-view/types"
