+++ apps/zui/src/zui-kit/react/list-view.tsx
@@ -13,7 +13,7 @@ import {createListView} from "../core/list-view/create-list-view"
 import {defaultListViewState} from "../core/list-view/state"
 import {ListViewArgs} from "../core/list-view/types"
 import {ReactAdapterProps} from "./types"
-import {useStateControllers} from "./use-state-controllers"
+import {useStateControllers} from "../utils/use-state-controllers"
 import {useInitialScrollPosition, useOnScroll} from "./utils"
 import classNames from "classnames"
 import {useParentSize} from "../utils/use-parent-size"
