+++ apps/zui/src/zui-kit/table/cell-component.tsx
