+++ apps/zui/src/zui-kit/table/config.ts
