+++ apps/zui/src/zui-kit/table/create-columns.tsx
@@ -1,5 +1,5 @@
 import * as zed from "@brimdata/zed-js"
-import {TableViewApi} from "src/zui-kit/core/table-view/table-view-api"
+import {TableViewApi} from "./table-view-api"
 import {ZedColumn} from "./column"
 
 export function createColumns(
