+++ apps/zui/src/zui-kit/table/grid.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import {VariableSizeGrid} from "react-window"
 import {Cell} from "./cell-component"
 import {useZedTable} from "./context"
-import {FillFlexParent} from "../fill-flex-parent"
+import {FillFlexParent} from "../utils/fill-flex-parent"
 import {InnerElement} from "./grid-container"
 import {useResizingClasses} from "./utils"
 import {call} from "src/util/call"
