+++ apps/zui/src/zui-kit/table/table-view-api.ts
@@ -1,17 +1,17 @@
 import {Table, createTable} from "@tanstack/table-core"
 import * as zed from "@brimdata/zed-js"
-import {config} from "../../../components/zed-table/config"
-import {Cell} from "../../../components/zed-table/cell"
-import {Position} from "../../../components/zed-table/position"
-import {getMaxCellSizes} from "../../../components/zed-table/utils"
-import {GridState, TableEvent} from "../../../components/zed-table/types"
-import {createColumns} from "../../../components/zed-table/create-columns"
-import {ZedColumn} from "../../../components/zed-table/column"
-import {TableViewArgs} from "./types"
+import {config} from "./config"
+import {Cell} from "./cell"
+import {Position} from "./position"
+import {getMaxCellSizes} from "./utils"
+import {GridState, TableEvent} from "./types"
+import {createColumns} from "./create-columns"
+import {ZedColumn} from "./column"
+import {TableViewArgs} from "../core/table-view/types"
 import {Controller} from "src/zui-kit/types/utils"
 import memoizeOne from "memoize-one"
-import {diffKeys} from "../utils/diff-keys"
-import {extractCellId} from "../utils/extract-cell-id"
+import {diffKeys} from "../core/utils/diff-keys"
+import {extractCellId} from "../core/utils/extract-cell-id"
 
 export class TableViewApi {
   element: HTMLDivElement | null = null
