+++ apps/zui/src/zui-kit/table/table-view.tsx
@@ -1,11 +1,11 @@
 import React, {forwardRef, useImperativeHandle, useMemo} from "react"
-import {Provider} from "../table/context"
-import {Grid} from "../table/grid"
+import {Provider} from "./context"
+import {Grid} from "./grid"
 import classNames from "classnames"
 import {TableViewArgs} from "../core/table-view/types"
-import {ReactAdapterProps} from "./types"
+import {ReactAdapterProps} from "../react/types"
 import {defaultTableViewState} from "../core/table-view/state"
-import {TableViewApi} from "../table/table-view-api"
+import {TableViewApi} from "./table-view-api"
 import {useStateControllers} from "../utils/use-state-controllers"
 import * as zed from "@brimdata/zed-js"
 
@@ -37,7 +37,6 @@ export const TableView = forwardRef(function TableView(
       shape,
       args.columnVisibleState.value,
       args.columnExpandedState.value,
-      ...args.deps,
     ]
   )
